+++ R/NAMESPACE.R
@@ -11,34 +11,29 @@
 #' @import rstantools
 #' @import shiny
 #' @import shinythemes
-#' @importFrom colourpicker colourInput 
+#' @importFrom ChangeR changePointsServer changePointsUI
 #' @importFrom DataTools checkAnyNonNumericColumns downloadModelUI downloadModelServer 
 #'  importDataUI importDataServer importUI importServer importOptions renameExistingNames
-#' @importFrom dplyr arrange bind_cols bind_rows cur_group_id distinct do group_by mutate n select
-#'  slice ungroup
+#' @importFrom dplyr %>% arrange bind_cols bind_rows cur_group_id distinct do group_by mutate n 
+#'  select slice ungroup
 #' @importFrom ggplot2 aes coord_cartesian element_line element_text 
 #'  geom_line geom_point geom_ribbon geom_vline ggplot ggplot_build ggtitle labs    
 #'  scale_colour_manual scale_fill_manual scale_shape_manual scale_size_manual 
 #'  scale_x_continuous scale_y_continuous sec_axis theme
-#' @importFrom htmltools save_html
-#' @importFrom jsonlite toJSON
-#' @importFrom loo loo loo_compare waic
-#' @importFrom magrittr %>%
-#' @importFrom mcp mcp
-#' @importFrom openxlsx write.xlsx
 #' @importFrom parallel detectCores
 #' @importFrom rlang .data
 #' @importFrom rstan sampling extract
 #' @importFrom shinycssloaders withSpinner
 #' @importFrom shinyjs alert enable runjs
-#' @importFrom shinyTools calculateRescalingFactors dataExportButton dataExportServer extractTitle
-#'  formatLegendOfGGplot formatPointsOfGGplot formatScalesOfGGplot formatTitlesOfGGplot
+#' @importFrom shinyTools addCustomPointsToGGplot calculateRescalingFactors
+#'  customPointsServer customPointsUI dataExportButton dataExportServer 
+#'  extractTitle formatPointsOfGGplot formatScalesOfGGplot formatTitlesOfGGplot
 #'  plotExportButton plotExportServer plotLegendServer plotLegendUI plotPointsServer plotPointsUI
-#'  plotRangesServer plotRangesUI plotTitlesServer plotTitlesUI
+#'  plotRangesServer plotRangesUI plotTitlesServer plotTitlesUI setLegendThemeOfGGplot
 #'  shinyTryCatch
 #' @importFrom shinyWidgets pickerInput updatePickerInput
-#' @importFrom stats approx as.formula dnorm lm median na.omit quantile sd
-#' @importFrom utils read.csv write.csv write.table combn
+#' @importFrom stats approx dnorm lm median na.omit quantile sd
+#' @importFrom utils write.csv combn
 #' @importFrom yaml read_yaml
 #' @references
 #' Stan Development Team (NA). RStan: the R interface to Stan. R package version NA. http://mc-stan.org